fix : 물품등록 관련 authority 수정
